Tests/LibWeb/Text/expected/HTML/set-outerHTML.txt

oldBox='<p>A box.</p>'
newBox='<p>Changed box.</p>'
NoModificationAllowedError: Cannot set outer HTML on document
